Fix to rita exit preventing possible infinite payments from client to exit
There is an issue within rita client that could result in an excess
payment since rita client pays the amount the exit gives even if the
local computation is different.

This combined with how the payments could take a while to verify means that the
client will continually send payments over to the exit which piles ontop
of itself, possibly resulting in a nasty feedback loop. This code will attempt to
subtract the excess and prevent this behavior.