Prevent repeated payments to rita exit
Althea is a pay per forward system, meaning everyone is paid for
forwarding traffic, this creates an obvious edge case, download traffic
which the user needs to pay for but is forwarded to them. The exit
resolves this issue by communicating to the client how much it owes, the
clients local computation can never be completely accurate due to the
potential of packetloss.

The issue arises since clients pay when their local debt exceeds a given
amount, since the client is just taking the debt value from the exit we
get the following edge case.

If a client sends a payment and the exit does not verify that payment
before the clietn gets around to checking if it should pay again, the
client may pay twice. If this delay is caused by the exit being
generally slow about verifying paymetns this may result in a feedback
loop, where slow payment verification inspires extra payments which
further slow payment verification.

This patch attemts to resolve the issue by subtracting the value of any
in flight payments from the outstanding debt of the user, but only when
we are sending the value to the client router. In this way the exits
internal interpretation of the users debt has not changed at all, but
the user will not pay twice because they will see their outstanding
payment already applied to the debt.

Co-authored-by: Justin Kilpatrick <justin@althea.net>