Restart Rita and babeld after an opkg update
Previously the old update.ash rebooted the router after updating packages, the beta 18 opkg update did not, resulting in a router that was updated,
but still running old software until an operator sent a manual reboot command.
This change restarts just Rita and babel after the opkg update, skipping the manual update/reboot which is what is currently being done.