Modify manual peers route to return if a modification is made
I don't really like the mutable reference pattern we've got going here
but tearing it out is pretty disruptive. In order to optimize our use of
the settings lock this function now returns true if it actually makes a
change, allowing us to skip updating the settings if they have not been
modified.