Peer discovery functions should not be members of TunnelManager
In the Actix Actor paradigm the Tunnel manager struct contained all the
state of the TunnelManager actor and any function that needed to
interact with this state had to be a member function. Now that we have
moved to async/await and our TunnelManager struct is held behind a
simple lock rather than the Actor message queue this results in a
deadlock.

Previously when we recieved a response from a hello message we would
send a call back message, now we can perform the callback within the
same async call stack as sending the hello if the first place. Since
sinding the hello requires access to the ports list, it was a
TunnelManager member function. This results in an impossible to resolve
deadlock where sending and recieving http hellos can not function.

This patch resolves the issue by removing the peer discovery functions
as TunnelManager member functions, resolving the deadlock.

We also move these peer discovery functions into their own file and make
a slight modification to the to_socket_addrs() use to ensure domains
resolve.