Modify integration tests to work with exits rename
In order not to conflict with the existing exits field in the config we
have to 'rename' the field as new_exits, this means our /settings
endpoint that merges raw json with the config does not work the same way
in the integration tests. This patch changes the integration tests to
use the new_exits field name.

This can be reverted once we finish removing the existing 'exits' field
which will be done automatically the first time a router saves the
config. From there we can use 'alias' in the following release and roll
back this change.