Bundle openssl by default
Recently due to cross compile issues in our package repo we moved to
bundling openssl for our openwrt release builds. As it turns out this
doesn't really increase binary size as much as we thought. In fact the
images end up slightly smaller.

This may be becuase what's added is easily compressed by squashfs or
becuase we're negating some advantages from the recent removal or legacy
futures and it's dependency tree. But for the time being we're moving to
bundled openssl in order to unify our cross compile config.