Optimizing writes to disk in order to prevent chip burnout
Previously, the various writes to disk would occur randomly
for usage tracker, debt keeper, and settings to the disk.
This isn't necessarily an issue, but it was very hard to ke-
ep track of when things were being saved and where they were
in the code respectively.

The code was all imported into a singular class for better clar-
ity as well as unifiying their constants to make the code easier
to understand. It was also changed to consider how much storage the
device had/if it was an exit as to better consider how often the
code should be saving to prevent oversaving if it didn't have much
storage.