Do not update balance if blockheight has been rolled back
In web30 0.18.3 we introduced a feature to check if a node is syncing
for balance checks.

The goal of this change was to get around issues where xdai nodes would
return old balances as they where syncing. It seems to have worked fine
during testing but this weekend we had an incident where many nodes
thought their balance was 0 due to a syncing node in the cluster. I
believe that this node must have been lying about syncing, returning
false when the eth_getSyncing endpoint was queried while it was in fact
very far behind the tip of the chain.

This patch adds another layer of protection, where the oracle keeps
track of the latest block it has seen and will refuse to update when
presented with older data. hopefully this is finally sufficient to
prevent issues.