Remove wait for settings
This two minute wait was added before the cron job based auto-rescue was
a thing so if we didn't wait long enough the router would simply fail to
work.

The two minute time was tuned specifically for the n600 and n750 devices
and doesn't hold true generally. Some devices like the glb1300 take much
longer than that to fully finish flashing.

But when you have an invalid config it's a two minute wait where it
seems like rita is working before it finally tells you the config is
invalid.

This patch removes the two minute wait, we'll rely on the cron job when
needed.