Add error return to handle_release_feed_update
This patch adds a result type to handle_release_feed_update, currently
if this does not work we attempt to opkg update anyways. It fails,
because the feed is not set correctly and even worse the logs are wiped
out when rita is rebooted at the end of the update process.

Now that we can handle the error here we will at least get the logs and
avoid disrupting the user for an update that we already know is doomed
to fail.