Fix opkg feed setting for new feed name
In the althea-firmware patch 'Add exit build support' the default feed
name for the althea package feed was changed from 'openwrt_althea' to
'althea' this breaks the logic that we use to avoid clobbering other
custom feeds as it was matching on openwrt_althea specifically.

Thankfully the fix is simple, we can just match althea and we will
successfully update old entires using openwrt_althea and fix opkg
feed updates on routers flashing with beta19 and later.