Edit wifi security on iface not device
Wireless security is a parameter of the connection not the device
because it's a logical software concept and does not effect how the
signal is transmitted in the real world. Which seems to be the divide
between these two sections of the config.

This patch modifies the wireless param on the correct section.