Display wifi encryption using readable names
This patch displays the wifi encryption modes using their maketing
designations rather than technical ones. In order to do this we provide
an alternate serialization for the EncryptionModes enum, allowing it to
be transformed into either human readable or technical desigation on the
fly.

We then have to use this struct when returning the wifi settings from
uci in order to matinain consistency, this requires serde-tagging and
special serialization/deserialization functions.

Finally we have to add a 'none' encryption mode as an option becuase
we're now parsing all wifi networks, including the mesh and phone wifi
networks that are disabled and hidden from the user by default. This
none encryption is exposed to the user as an option as well.