Fix router time set;  move time sync to its own file;
Recent optimizations to reduce the amount of disk wirtes on the rotuers
seem to have resulted in clocks becoming less accurate. Right now we
assume this is because the router time is set to the latest file written
time for lack of a better default. So file write optimizations result in
routers with clocks that are days behind.

This causes specific issues with wireguard where it will only accept a
handshake in the future once it already hs a handshake from a router.
For a router on first boot this is no issue, any time is later than 0
but if a router reboots, now has a time a day old, it will never
successfully re-handshake with the exit.

This patch resolves the issue by adding an endpoint to the exit which
serves the current unix time and functionality to the router to query
this exit when the wg_exit tunnel has not had a successful handshake
recently, providing a recovery path from this edge case.

Commit message written by Justin Kilpatrick