Handle non-zero return codes from opkg
Previously the update system function only handled kernel interface
errors. For example if the opkg binary did not exist, or was corrupted
an error would be returned. But if opkg ran and returned an error this
was not detected because the return value would be Ok(process output).

This patch looks at the process output return code and uses that to
determine if the update has actually succeded.

We also move to logging the full output of commands in every state as
updates are rare the logging load is low and of high value.

Some testing was done with opkg to determine that it returns a zero exit
code even if warning messages are printed as long as it has performed
the main task required of the command.