Parsing json/bincode files to print out usage data
This code takes usage tracker file and removes a lot of the unnecessary
logic and attempts to parse a tarred file containing bincode that stores
usage data information. This is a binary executable that attempts to format
the data into a csv file that consists of m/y - amount and splits lines based
on device.