Setup endpoint on exit that services a list of exit ips in a cluster
Previously clients assume that each exit is a single identity, with a single ethereum key and wg_key combo.
this patch will handle the other two components of an identity. This is needed for exit switching so that
wireguard interfaces dont recieve keepalives from several exits instances resulting in route flapping