Add functions for grabbing email and phone
The ContactType struct was designed to handle very strictly the possibly
combinations of email and phone number information to ensure that we can
always contact the customer as effectively as possible. That being said
the struct enum has proven to be unwieldy to work with in a lot of ways.

This patch adds some easy functions for grabbing the phone number and
email which should dramatically cut down on the amount of boilerplate
logic used when handling this struct.