Remove uci for setting slaac config
Prebiously uci was used to setup an ipv6 subnet in slaac, which required a /etc/init.d/network reload. Doing
this reload would cause other addresses added to br-lan to get cleared which was problem in lte routers which required a
subnet on br-lan. This commit changes this by manually setting up an address on br-lan