Fix ssh keyfile permissions and increase robustness
This set of changes fixes a key issue with the ssh key insertion
operator action. Previously the new file was created with inappropriate
permissions locking users out of the router immediately after it was
used.

In addition to this permissions change the code has been re-organized to
be more robust to errors, including creating the keyfile if it does not
already exist for some reason and gracefully recovering from errors that
may occur while creating the new file, changing it's permissions, or
swapping it out.