Do not classify unenforced traffic
The act of classifying traffic that is not being enforced by the exit
slows it down due to the buket process used by htb. In order to avoid
this we will now delete traffic classes when users are not enforced.

Also modified in this patch, we no longer specify the burst and quantum
and instead allow them to be computed by HTB based on the r2q and target
rate.