Independent operator update loop
This patch isolates the operator update loop from the rita client loop,
the goal here is to ensure that no issue in the main rita client loop
causes the router to stop accepting instructions from operator tools.
This is key becuase we depend on operator updates to send things like
reboot commands, key for helping clients recover and totally useless if
the operator update loop is locked up.