Set mtu of wg_exit on exit side to 1500
Prebiously this was set to 1340. It was noticed that the mtu on ex nic of the exit was 1500
When large packets went out, they were fragmented at wg_exits mtu of 1340, But when large packets came in
they were fragmented at the ex nic mtu of 1500, meaning they could not get forwarded to wg_exit and a Packet
too big message was sent. Increasing mtu on the exit fixes this issue