Setup iptables and ip route after wg tunnel has been setup
Previously these were being setup before the wg_exit tunnel was setup, which caused unneccessary downtime since they
needed the tunnel to be setup to setup the routes. This wasted a loop and lasted a couple of minutes. Now we setup the
tunnel first and then setup routes and rules for all clients in question