Exits advertises two mesh ip addrs
In continuation of setting up exit roaming and switching, in order to add backward compatibility between b19 and b20 routers,
two wg exit interfaces were added. This commit continues that by getting the exit to advertise two ip addrs, one for b19 routers to connect to
as normal, and one for b20 routers to allow for exit roaming
B20 routers retrieve a list of IP addrs from an exit and uses these ips to migrate between exits. Since all b19 routers are connected on a seperate ip
this should not affect them.