Fix binary opt
An unintentional change snuck in reducing this to just '2' level opt
while I was trying to speed up debugging. The impact on openwrt builds
does not seem to have been too bad. But we should build optimized
binaries by default to eek out as much storage space as we can.