Don't panic on babel connect_timeout
This is a bit of a smoking gun when it comes to why we're having so many
babel related errors. If for some reason babel is not ready to connect
within a fairly short timeout we panic the entire thread making the
call. This change alone will help immensely with babel resilency.

This along with several other spurious panics have been resolved.