Auto restart babel in case of failure
This change will automatically restart babeld ~11 minutes after it
crashes or otherwise starts acting strangely.

We count not being able to connect to babel, not being able to parse
it's output and not being able to run commands as failures, after enough
consecutive failures pile up a babel restart is initiated.