Refine exit registration state machine
This patch removes two unneeded states from the exit registration state
machine and also parallelizes the status requests.

The exit state removal removes an unused state and an invalid state only
used on the exit which would result in routers not continuing to make
rqeuests and effectively going dark.

The parallelization reduces the time for an avearge tick of the client
loop from 30s to 5s