Add make payments v2
This new endpoint takes a list of transactions instead of just one, every payments we make, we send
over this list and the recepient verifies this list with its own list and sends the difference to
payment validatot
This allows endpoints to synchronise transaction, credit missing transactions, and account for
transcations that were missed.
We also send this list to all exits instead of just the one we are connected to. This allows us
ensure that payments work in the case of exit switching. In the worst case, client get double credited.
This will last until exits have different payment addresses