Create make_payment_v2
This new endpoint takes a list of transactions instead of just one.
All nodes store all TXs sent from startup. For every payments a client makes make, they send
over this list and the recepient verifies this list with its own list and sends the difference to
payment validator
This allows endpoints to synchronise transaction and account for
transcations that were missed.