Add Rita Extender
We expand the lan of the primary althea routers using a new extender
version of the firmware that can either be connected via wirelessly or with a wire.

A new crate for extender is added. A new endpoint is setup on rita client
Extender queries this endpoint to get a list wifi settings, and applies those same settings
to its wifi devices