Fix operator_fee display and behavior
This fixes the /prices API endpoint field, and adds the ability to save an operator_fee change.  This also fixes a bug that previous would overwrite when an op fee was set higher than the default.  This now allows that, but will replace an op fee that's lower than the default.