Increase recusion limit in write_all_spinlock
Previously the recusion limit for this function was 100, this function
was refusing to move a 1.2MB CSS file becuase it would hit over 100
recusions before succeeding. Considering each packet is 1340 bytes this
isn't suprising. Assuming one packet per loop that's a max of 134kb but
presumably we often get to read and write more than one packet at a
time so it took a much larger file to discover that this limit was
problematic.

What this patch does not fix fundamental architecture issues in antenna
forwarder that make forwarding large files slow. Mainly the bottleneck
between the server and the client along a single stream. I suspect
changing the send and recieve buffer size for the sockets may help a
lot.