Set tcp nodelay in write_all_spinlock
This change sets the tcp nodelay flag on sockets before we write data.
this means that instead of waiting for 'enough' data to build up on the
socket it will be sent right away. While this does seem to have a
moderate performance improvement the real issue is that the default
tcp socket buffer size is small. This means when we get a large file
from say the client or the antenna we quickly read this data out off of
the socket and into our own buffer and then slowly meter it into the
outgoing socket. Likewise the other end experiences this slow drip of
data as a low download speed.

This minor improvement is worth comitting but is not the correct long
term solution.