Optimise startup setup time for rita exit
Setup clients takes ~5 to complete on the first tick since it is settings up routes and rules for all
clients in the database. This commit modifies so it sets ups routes and rules only for clients that have an
active handshake value. There are over 1000 clients, but only around 200 actively handshaking, so this should
reduce startup time dramtically