Fix poorly designed interface contains check
This patch updates the routine that checks if we have a wg interface for
a specific wg number. Previously this routine simply took the string
output of 'ip link' and checked if it contained a given wg name.

This is a terrible idea for a number of reasons, lets say you had an
interface wg10 which was created and later deleted. Interface wg100
would then block the creation of a new wg10 becuase 'wg10' has a match.

This is only one of the ways this could go wrong off the top of my head.

A very simple change instead gets a list of interfaces and does a more
exact string comparison.