Fix ipv6 proc parsing
This parser implementation was flawed, resulting in a collision when
parsing interface names common on 32x routers where lan1 and wlan1 are
both valid interfaces. Since the function returned the wrong ip peer
listening would always fail on that interface.

This patch fixes the parser and also improves the test code to be more
reliable.