Logic erorr in Fix: batch payment verifications
This patch fixes a logic error in the recent comment "Fix: batch payment
verifications". The last few futures in the buffer would never execute
resulting in payments not being verified. For example if 11 payments
where sent the first 10 would be verified by the last 1 would be
discarded by accident.

This patch corrects that logic issue by processing the futures left in
the buffer.