Reduce exit routes
This patch reduces the number of exit routes installed by a factor of
several hundred, bringing us down from 4k ipv6 routes to a much more
manageable 27 and maintaining the full functionality.

Similarly hundreds of ipv4 routes are eliminated by only assigning ipv4
internal routes to wg_exit users and leaving the default route for the
subnet on wg_exit_v2. Since there are only a few wg_exit users left this
brings us from ~550 ip4 routes to ~20

Also added as an option in this patch is a way to disable the root htb
qdisc to improve routing speed temporarily.