Add client ipv6 to wg_exit
We assign this routers ipv6 subnet to br-lan so that it is used for slacc address assignments to clients. But this created a
problem with default routes, since there was no ipv6 address in the br-lan subnet on wg_exit (the interface that the ipv6 default route is on) ipv6 traffic would be hit or miss.
In some cases by methods unknown to us the source ip would be correct and ipv6 would work, in other cases it would not.

This patch resolves the issue by assigning the first ip in the routers subnet to wg_exit as a /128 in addition to putting
the entire subnet on br-lan.
This means that the first ip in the subnet is the routers ip on both lan and wg_exit, correcting the source address and resulting in more reliable ipv6 operation