Add blockchains to test-runner, refactor scripts
This patch heavily refactors the scripts directory in this repo as well
as the new integration test, renaming the old integration test the as
'legacy' and setting up the new test environment for it's primary goal,
The ability to execute many distinct scenarios and run with an actual
blockchian backend from a pure rust coordination environment.

The pure rust requirement will allow us to more easily describe complex
network arch as well as scale up testing with dramatically more nodes
than was previously possible.