Modfiy root htb parameters
HTB is the root qdisc we use for the exits to enforce on specific
classes of traffic (we filter on the internal ip of enforced routers
which is itself enforced by wireguard which will not allow that client
to communicate over a different ip)

Sadly htb is fundamentally single threaded, meaning it's the singular
performance and scaling bottleneck for our exit operations currently. As
packets must go through the classification filter before they can be
handed off to other parts of the kernel which successfully handle
encryption and forwarding in parallel.

What this patch does will not fix the fundamental issue. What it does is
modify two parameters the direct_qlen and the r2q value. What both of
these relate to is how many packets are forwarded per classification
operation before the single threaded htb logic is run again.

In theory the idealize throughput will be doubled by these changes since
the r2q value will now forward twice as many packets per classification.
The impact of the increased qlen is harder to estimate. In my current
mental model (which may be incorrect) traffic is classified by htb, then
queued for each class, so how many packets can go through per
classification operation is the minimum of the r2q computed quantum
(number of bytes to forward per classification) and the remaining space
on this queue. So the ideal value is just big enough that remaining
space on the queue is never an issue, figuring out if this actually is
sufficient would require more careful inspection of performance
counters.