Update dns servers provided by dhcp
This patch adds some code to rita client that updates the dns server
list provided by the internal router dhcp server to clients.

This is important because so far we have been handing out 1.1.1.1 and
8.8.8.8 as the first servers on this list. While modern linux resolvers
will round robin and fallback to later entires some devices do not.
Meaning we have users in the field who simply see dns errors when their
devices does not get a response from the first server on this list.

There are of course many devices that go out to their own servers as
fallbacks or can otherwise recover from these errors.

This change helps handle those less robust devices by directing their
queries to the local router, who will then forward them to the exit
internal dns server primarily, but fallback with other servers in the
routers resolve.conf list if that server is down.