Fix: Update dns servers provided by dhcp
As it turns out this commit was completely misguided, the dnsmasq server
config attribute decides what servers are used for upstream resoltuion
(as in what dnsmasq uses to resolve requests). So setting the value to
192.168.10.1 results in an infiniate loop of a request.

This patch modifies the logic to insert the exit ip address to the front
of any resolver list that is found in dnsmasq, this should then match
what is in the current default firmware configuration.