Fix: Potential index out of bounds
As a classic example of why using typesafe operations should be required
even for people who 'know what they are doing'.

When I removed the first check in the previous commit for a logic change
I made it so that this code would panic if there was no server entry.
When in fact we want that condition to pass unremarked.