Fix: Validate payment with Usage History before adding pmt
This is one of the classic blunders, trying to edit an array under
yourself by having multiple copies of it and then becoming confused as
the indexes no longer line up.

The patch fixes modification issue by building a new array with only
valid data and simply dropping invalid data instead of attempting to
remove it.