Testrunner Exit support
This adds support for setting up exits to the testrunner. The
architecture here is to have exits with actual access to the internet
via a bridge interface, the postgresql db is setup entierly in Rust and
is located in the native namespace, meaning the testrunner itself has
full database access to make experimental changes.

Unlike the previous legacy integration tests which supported only one exit this
version supports an arbitrary number of exits which will allow for
testing of failover scenarios as well.