Remove recursion from antenna forwarder spinlock
In very strange connection cases it is valid for this spinlock to run
for a very long time, but we can't recurse that many times without
overflowing our stack. This patch modifies the structure of
write_all_spinlock to be iterative rather than recursive so that it can
operate correctly even during very strange conditions.