Only add enforcmenet classes when needed
This patch moves setup of enforcment filter classes from startup to
runtime where they are only added if a client is being enforced upon.

This change does not remove filters when enforcement is done, which
isn't optimal, but the main performance optimization is to reduce the
number of filters added overall by only setting them up for active
enforced users, especially since we do not add filters for every single
potential ip a client has and instead only for the currently active ipv6
network if any.