Eliminate lazy_static for exit wg_usage
This patch eliminates the use of a lazy static global reference for the
exit wg traffic counting. This is a strictly superior solution since we
don't need to access the usage reference outside of the context of the
rita exit loop.

Lazy static effectively makes the reference to a given lock a global
variable, this causes issues in our test environment where we want to
run multiple threads of the exit in the same context since the
idependent instances of the code cross-talk through the lazy static
variables.