Remove lazy_static rita_exit settings cache
This was added long ago when the exit was having deadlock issues with
the settings, a problem that has been fundamentally solved with the
settings refactor where we now get cloned copies of the settings in
essentially all cases and never hold the lock. Since that factor is
handled we can safely remove these locks and reduce code complexity.