Settings + TunnelManager namespace handling
Previously the feature 'read from file' was used in the settings crate
to handle the testrunner issue where different instances of rita being
spawned under one rust process would read each others configs.

This patch pushes a new solution to that problem, since all lazy static
structs are subject to the issue and the 'read from file' method in the
settings crate was not a portable solution.

The lazy static getter and setter functions for settings and tunnel
manager have been updated to be net namespace aware at least when kernel
interface is built with the correct feature. They will appropriately
dispatch requests to a different copy of the state based the net
namespace.