Force Althea addresses derived from ID eth key
This patch removes routers from having to manage both an Althea address
and a ethereum address. Having to values was confusing enough and we
where already forcing the private key to be the same on the actual
router itself.

But transmitting the althea address and eth address as distinct fields
still provided a lot of room for bugs and problems. Namely even though
we programmed the routers to have althea address and eth address be the
same public key, there's nothing actually enforcing that future
implementations do so.

This patch reverts the addition to the identify field and in doing so
enforces that nodes have one key and that they althea key is to be
derived from the eth key.